+++ ecs/systems/s_camera.cpp
@@ -11,17 +11,19 @@
 #include "game.h"
 
 #define EPSILON 0.1
-#define DAMPNESS 0.05 //Lower is slower
 
 using namespace Verse;
 
 namespace {
     glm::mat4 mat_active_camera;
     Component::Camera *cam; //Active Camera
+    Rect bounds;
 
     Vec2 focus_pos;
     float look_ahead = 32;
+    float la_speed = 1;
     float curr_la, target_la;
+    bool is_la_stopped;
 }
 
 void System::Camera::init(Component::Camera* camera, Vec2 pos, Vec2 size) {
@@ -36,6 +38,12 @@ void System::Camera::init(Component::Camera* camera, Vec2 pos, Vec2 size) {
 void System::Camera::setActive(Component::Camera *camera) {
     Graphics::Renderer::GL::useCamera(&mat_active_camera, &camera->centre);
     cam = camera;
+    bounds = Rect(0,0,0,0);
+}
+
+void System::Camera::setActive(Component::Camera *camera, Rect cam_bounds) {
+    setActive(camera);
+    bounds = cam_bounds;
 }
 
 void System::Camera::update(Config &c, Scene &scene) {
@@ -55,19 +63,29 @@ void System::Camera::update(Config &c, Scene &scene) {
 void System::Camera::move(Config &c, Vec2 pos, int input) {
     updatePoints(c, pos);
     
-    if (cam->vel.x != 0) {
-        target_la = -sign(cam->vel.x) * look_ahead;
+    is_la_stopped = (input != 0 and input == sign(cam->vel.x)) ? false : is_la_stopped;
+    
+    if (not is_la_stopped) {
+        if (input == 0) {
+            is_la_stopped = true;
+            target_la = curr_la + (sign(cam->vel.x) * look_ahead - curr_la) * 0.25;
+        } else if (cam->vel.x != 0) {
+            target_la = sign(cam->vel.x) * look_ahead;
+        }
     }
     
-    //Smooth transition
     if (abs(target_la - curr_la) > EPSILON) {
-        curr_la += (target_la - curr_la) * DAMPNESS;
+        curr_la += la_speed * sign(cam->vel.x) * ((target_la - curr_la) / target_la);
     }
     
-    cam->centre = focus_pos + Vec2(curr_la, 0);
+    focus_pos.x += curr_la;
     
-    mat_active_camera = glm::translate(glm::mat4(1.0f), glm::vec3((cam->centre.x - 0.5 * c.resolution.x) / c.render_scale,
-                                                                  (cam->centre.y - 0.5 * c.resolution.y) / c.render_scale, 0.0f));
+    checkBounds(c);
+    
+    cam->centre = focus_pos;
+    
+    mat_active_camera = glm::translate(glm::mat4(1.0f), glm::vec3((0.5f * c.resolution.x - cam->centre.x) / c.render_scale,
+                                                                  (0.5f * c.resolution.y - cam->centre.y) / c.render_scale, 0.0f));
 }
 
 void System::Camera::updatePoints(Config &c, Vec2 &pos) {
@@ -87,5 +105,25 @@ void System::Camera::updatePoints(Config &c, Vec2 &pos) {
     cam->t += cam->vel.y;
     cam->b += cam->vel.y;
     
-    focus_pos = Vec2(floor(c.resolution.x - (cam->l+cam->r)/2.0f), floor(c.resolution.y - (cam->t+cam->b)/2.0f));
+    focus_pos = Vec2(floor((cam->l+cam->r)/2.0f), floor((cam->t+cam->b)/2.0f));
+}
+
+void System::Camera::checkBounds(Config &c) {
+    if (bounds.size.x != 0) {
+        if (focus_pos.x + (0.5f * c.resolution.x) > bounds.pos.x + 0.5f * bounds.size.x) {
+            focus_pos.x = bounds.pos.x + 0.5f * bounds.size.x - 0.5f * c.resolution.x;
+        }
+        if (focus_pos.x - (0.5f * c.resolution.x) < bounds.pos.x - 0.5f * bounds.size.x) {
+            focus_pos.x = bounds.pos.x - 0.5f * bounds.size.x + 0.5f * c.resolution.x;
+        }
+    }
+    
+    if (bounds.size.y != 0) {
+        if (focus_pos.y + (0.5f * c.resolution.y) > bounds.pos.y + 0.5f * bounds.size.y) {
+            focus_pos.y = bounds.pos.y + 0.5f * bounds.size.y - 0.5f * c.resolution.y;
+        }
+        if (focus_pos.y - (0.5f * c.resolution.y) < bounds.pos.y - 0.5f * bounds.size.y) {
+            focus_pos.y = bounds.pos.y - 0.5f * bounds.size.y + 0.5f * c.resolution.y;
+        }
+    }
 }