+++ ecs/systems/s_camera.h
@@ -15,7 +15,9 @@ namespace Verse::System::Camera
 {
     void init(Component::Camera* camera, Vec2 pos, Vec2 size);
     void setActive(Component::Camera* camera);
+    void setActive(Component::Camera* camera, Rect cam_bounds);
     void update(Config &c, Scene& scene);
     void move(Config &c, Vec2 pos, int input);
     void updatePoints(Config &c, Vec2 &pos);
+    void checkBounds(Config &c);
 }